deleting helper module and moving while_connected class to esl module
