+++ greenswitch/esl.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import errno
+import functools
 import logging
 import pprint
 import sys
@@ -467,6 +468,42 @@ class OutboundSession(ESLProtocol):
         """
         if not self._outbound_connected:
             raise OutboundSessionHasGoneAway
+
+    def while_connected(self):
+        """Returns a object that check if the session is connected in
+        the __enter__ and __exit__ steps, if not connected will
+        raise the greenswitch.esl.OutboundSessionHasGoneAway exception.
+
+        This method can be used as a context manager or decorator.
+
+        Examples:
+        >>> with outbound_session.while_connected():
+        >>>    do_something()
+        >>>
+        >>> @outbound_session.while_connected()
+        >>> def do_something():
+        >>>     ...
+        """
+        class _while_connected(object):
+            def __init__(self, outbound_session):
+                self.outbound_session = outbound_session
+
+            def __enter__(self):
+                self.outbound_session.raise_if_disconnected()
+                return self
+
+            def __exit__(self, exit_type, exit_value, exit_traceback):
+                self.outbound_session.raise_if_disconnected()
+
+            def __call__(self, func):
+                @functools.wraps(func)
+                def decorator(*args, **kwargs):
+                    with self:
+                        return func(*args, **kwargs)
+
+                return decorator
+
+        return _while_connected(self)
 
 
 class OutboundESLServer(object):
